Updates to DiskTensor chunking abilities.
Having the data properly chunked to mirror access patterns greatly improves the
performance of the code. For example, one test with the old chunking code
took 15-16 seconds to perform the write. With proper chunking the write
dropped to 1.2 seconds. Having read access equivalent to the chunking
also greatly improves performance.

I also removed the default compression of the data as it was a measurable
performance penalty.