FEAT: Add HPTT Unit tests.
FIX: Unit tests exposed bug when fusing indices containing offsets - this is now prevented.
FIX: Stride order check sign corrected. Asserts that the next stride cannot be greater than the current one.
FIX: When creating TensorViews smaller than the source dimensions, unused dimensions are concatenated to the inner not outer dimensions as this will cause offsets to lose their stride size otherwise. (i.e. in Tensor A(4, 5, 3, 6) with strides (90, 18, 6, 1) the TensorView A_View(2, 2) with offsets (1, 1) and strides (90, 6) need source dimenions of ([at least] 3, 15) so that HPTT knows to create strides of (90, 6) when provided with an inner stride of 6.